Update Order and Product models to use decimal for monetary values
- Changed Freight and UnitPrice properties from double? to decimal?
  in Order, OrderDetail, and Product classes for improved precision
  in financial calculations.

+++ Northwind.EntityModels.Sqlite/Order.cs
@@ -39,7 +39,7 @@ public partial class Order
     public int? ShipVia { get; set; }
 
     [Column(TypeName = "money")]
-    public double? Freight { get; set; }
+    public decimal? Freight { get; set; }
 
     [Column(TypeName = "nvarchar (40)")]
     [StringLength(40)]
